RAG_Production_APP/requirements.txt
fastapi
inngest
llama-index-core
llama-index-readers-file
python-dotenv
qdrant-client
uvicorn
streamlit
google-generativeai
langchain-google-genai
slack_sdk
